+++ Gitagram/View/PostView/PostView.swift
@@ -1,76 +1,117 @@
 //
-//  PostView.swift
+//  PostGaugeView.swift
 //  Gitagram
 //
-//  Created by saki on 2024/05/04.
+//  Created by saki on 2024/07/03.
 //
 
 import SwiftUI
+import Combine
 
 struct PostView: View {
-    @State var cardData: CardData = .Empty()
-    @State var title: String = ""
-    var gitHubLoader = GitHubLoader()
-    @StateObject  var viewModel = RepositoryViewModel()
-    
+    @State private var value: Double = 0.2
+    @State var selection = 1
+    @StateObject var postViewModel = PostViewModel()
+    let gadient = Gradient(colors: [.pink,.purple])
+    init() {
+        UITabBar.appearance().isHidden = true
+    }
     var body: some View {
-       
-            VStack{
-               
-                Text("リポジトリの名前は？")
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.leading,10)
-                    .font(.system(size: 30, weight: .black, design: .default))
-                    .padding(.bottom,30)
+        
+        VStack{
+            Gauge(value:value){
                 
-                TextField("リポジトリの名前を入力してね", text: $title)
-                    .frame(alignment: .leading)
-                    .padding(.leading,10)
-                    .onChange(of: title) {
-                        let product = cardData.product.setTitle(from: title)
-                        cardData = cardData.setProduct(from: product)
+            }
+            .padding(.bottom, 50)
+            .padding(.top, 80)
+            .tint(gadient)
+            
+            
+            TabView(selection: $selection) {
+                PostTitleView()
+                    .tag(1)
+                    .environmentObject(postViewModel)
+                    .onChange(of: selection) { newValue in
+                        if newValue == 1 {
+                            withAnimation{
+                                value = 0.2
+                            }
+                            print(postViewModel.cardData)
+                            
+                        }
                     }
                 
-                Divider()
-                
-                Text("または下のリストから選択")
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.leading,10)
-                    .padding(.top, 20)
-                    .font(.system(size: 20, weight: .black, design: .default))
-                    .padding(.bottom,5)
-                
-                
-                List(viewModel.repositories) { repository in
-                    Button(action: {
-                        title = repository.name
+                PostDescriptionView()
+                    .tag(2)
+                    .environmentObject(postViewModel)
+                    .onChange(of: selection) { newValue in
+                        if newValue == 2 {
+                            withAnimation{
+                                value = 0.4
+                            }
+                            print(postViewModel.cardData,"2")
+                        }
                         
-                    }, label: {
-                        Text(repository.name)
-                            .font(.headline)
+                    }
+                
+                PostURLView()
+                    .tag(3)
+                    .environmentObject(postViewModel)
+                    .onChange(of: selection) { newValue in
+                        if newValue == 3 {
+                            withAnimation{
+                                value = 0.6
+                            }
+                            print(postViewModel.cardData,"3")
+                        }
+                    }
+                
+                PostHashTagView()
+                    .tag(4)
+                    .environmentObject(postViewModel)
+                    .onChange(of: selection) { newValue in
+                        if newValue == 4 {
+                            withAnimation{
+                                value = 0.8
+                            }
+                            print(postViewModel.cardData,"3")
+                        }
                         
-                    })
-                }
+                    }
                 
-                Spacer()
-                
-          
+                PostImageView()
+                    .environmentObject(postViewModel)
+                    .tag(5)
+                    .onChange(of: selection) { newValue in
+                        if newValue == 5 {
+                            withAnimation{
+                                value = 1.0
+                            }
+                        }
+                    }
+            }
+            .tabViewStyle(PageTabViewStyle(indexDisplayMode: .never))
+            .tabViewStyle(.page)
             
-        }
-        .onAppear(){
-            Task{
-                if let host = await GetLoginDeveloperUseCase().execute() {
-                    cardData = cardData.setLoginHost(from: host)
-                    let username = cardData.loginHost.name
-                    viewModel.fetchRepositories(for: username)
-                } else {
-                    ContentView()
-                }
+            if selection != 5{
+                Button(action: {
+                    selection += 1
+                    print(postViewModel.cardData,"ボタン押されたよ")
+                }, label: {
+                    Text("次へ")
+                        .padding(.horizontal,120)
+                        .padding(.vertical,15)
+                        .font(.system(size: 10, weight: .medium, design: .default))
+                        .foregroundColor(.white)
+                        .background(Color(Color(red: 0.82, green: 0.6, blue: 0.97)))
+                        .cornerRadius(30)
+                        .padding(.bottom,20)
+                })
             }
         }
     }
+    
 }
-
 
 #Preview {
     PostView()
