+++ Gitagram/View/PostView/PostViewModel.swift
@@ -1,0 +1,11 @@
+//
+//  PostViewModel.swift
+//  Gitagram
+//
+//  Created by saki on 2024/07/04.
+//
+
+import Foundation
+class PostViewModel: ObservableObject {
+    @Published var cardData: CardData = .Empty()
+}
